Feat/Add support for modular arithmetic and EC add in the eDSL (#293)
* Add air.rs

* Add AIR constraints

* Add trace generation

* Correct range check for `total_quotient`

Still missing to fix `total_quotient` calculation in trace when `total` is negative

* Fix total_quotient constraint and trace generation

* Add passing test

* Add more tests

* Improve range check

* Optimize columns

* Optimize columns further

* Further durther optimize columns

* Disambiguate mod mul AIR

* Reorganize

* Refactor out common AIR constraints

* Refactor out common trace generation logic

* Add doc comment

* Finish bigint chip and tests

* Optimize parameters

* Add #[cfg(test)]

* Initial commit

* Commit

* Add compiler stuff

* Add conditionals for bigint

* EC double working

* Finish ec_add

* Add comments explaining constants

* chore: rename modules

* Remove dependency on primitives

* Rename variable

* Remove debugging code

* Rename stuff to include mod

* Rustfmt

---------

+++ primitives/src/modular_multiplication/bigint/trace.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use itertools::Itertools;
-use num_bigint::BigUint;
+use num_bigint_dig::BigUint;
 use p3_air::BaseAir;
 use p3_field::PrimeField64;
 use p3_matrix::dense::RowMajorMatrix;
@@ -44,8 +44,8 @@ impl<F: PrimeField64> LocalTraceInstructions<F> for ModularMultiplicationBigIntA
 
     fn generate_trace_row(&self, input: Self::LocalInput) -> Self::Cols<F> {
         let (a, b, range_checker) = input;
-        assert!(a.bits() <= self.total_bits as u64);
-        assert!(b.bits() <= self.total_bits as u64);
+        assert!(a.bits() <= self.total_bits);
+        assert!(b.bits() <= self.total_bits);
 
         let range_check = |bits: usize, value: usize| {
             let value = value as u32;
